+++ SpatialGDK/Source/SpatialGDK/Private/Utils/ComponentFactory.cpp
@@ -434,8 +434,7 @@ TArray<FWorkerComponentUpdate> ComponentFactory::CreateComponentUpdates(UObject*
 	// Only support Interest for Actors for now.
 	if (Object->IsA<AActor>() && bInterestHasChanged)
 	{
-		InterestFactory InterestUpdateFactory(Cast<AActor>(Object), Info, EntityId, NetDriver->ClassInfoManager, NetDriver->PackageMap);
-		ComponentUpdates.Add(InterestUpdateFactory.CreateInterestUpdate());
+		ComponentUpdates.Add(NetDriver->InterestFactory->CreateInterestUpdate((AActor*)Object, Info, EntityId));
 	}
 
 	return ComponentUpdates;