+++ SpatialGDK/Source/SpatialGDK/Public/Utils/InterestFactory.h
@@ -7,74 +7,119 @@
 
 #include <WorkerSDK/improbable/c_worker.h>
 
+/**
+ * The InterestFactory is responsible for creating spatial Interest component state and updates for a GDK game.
+ *
+ * It has two dependencies:
+ *   - the class info manager for finding level components and for creating user defined queries from ActorInterestComponents
+ *   - the package map, for finding unreal object references as part of creating AlwaysInterested constraints
+ *     (TODO) remove this dependency when/if we drop support for the AlwaysInterested constraint
+ *
+ * The interest factory is initialized within and has its lifecycle tied to the spatial net driver.
+ *
+ * There are two public types of functionality for this class.
+ *
+ * The first is actor interest. The factory takes information about an actor (the object, info and corresponding entity ID)
+ * and produces an interest data/update for that entity. This interest contains anything specific to that actor, such as self constraints
+ * for servers and clients, and if the actor is a player controller, the client worker's interest is also built for that actor.
+ *
+ * The other is server worker interest. Given a load balancing strategy, the factory will take the strategy's defined query constraint
+ * and produce an interest component to exist on the server's worker entity. This interest component contains the primary interest query made
+ * by that server worker.
+ */
+
 class UAbstractLBStrategy;
 class USpatialClassInfoManager;
 class USpatialPackageMapClient;
-class AActor;
 
 DECLARE_LOG_CATEGORY_EXTERN(LogInterestFactory, Log, All);
 
 namespace SpatialGDK
 {
+
 class SPATIALGDK_API InterestFactory
 {
 public:
-	InterestFactory(AActor* InActor, const FClassInfo& InInfo, const Worker_EntityId InEntityId, USpatialClassInfoManager* InClassInfoManager, USpatialPackageMapClient* InPackageMap);
+	InterestFactory(USpatialClassInfoManager* InClassInfoManager, USpatialPackageMapClient* InPackageMap);
 
-	static void CreateAndCacheInterestState(USpatialClassInfoManager* ClassInfoManager);
+	Worker_ComponentData CreateInterestData(AActor* InActor, const FClassInfo& InInfo, const Worker_EntityId InEntityId) const;
+	Worker_ComponentUpdate CreateInterestUpdate(AActor* InActor, const FClassInfo& InInfo, const Worker_EntityId InEntityId) const;
 
-	Worker_ComponentData CreateInterestData() const;
-	Worker_ComponentUpdate CreateInterestUpdate() const;
-
-	static Interest CreateServerWorkerInterest(const UAbstractLBStrategy* LBStrategy);
+	Interest CreateServerWorkerInterest(const UAbstractLBStrategy* LBStrategy);
 
 private:
+	// Shared constraints and result types are created at initialization and reused throughout the lifetime of the factory.
+	void CreateAndCacheInterestState();
+
 	// Build the checkout radius constraints for client workers
-	static QueryConstraint CreateClientCheckoutRadiusConstraint(USpatialClassInfoManager* ClassInfoManager);
-	static QueryConstraint CreateLegacyNetCullDistanceConstraint(USpatialClassInfoManager* ClassInfoManager);
-	static QueryConstraint CreateNetCullDistanceConstraint(USpatialClassInfoManager* ClassInfoManager);
-	static QueryConstraint CreateNetCullDistanceConstraintWithFrequency(USpatialClassInfoManager* ClassInfoManager);
+	// TODO: Pull out into checkout radius constraint utils
+	QueryConstraint CreateClientCheckoutRadiusConstraint(USpatialClassInfoManager* ClassInfoManager);
+	QueryConstraint CreateLegacyNetCullDistanceConstraint(USpatialClassInfoManager* ClassInfoManager);
+	QueryConstraint CreateNetCullDistanceConstraint(USpatialClassInfoManager* ClassInfoManager);
+	QueryConstraint CreateNetCullDistanceConstraintWithFrequency(USpatialClassInfoManager* ClassInfoManager);
 
 	// Builds the result types of necessary components for clients
-	static TArray<Worker_ComponentId> CreateClientNonAuthInterestResultType(USpatialClassInfoManager* ClassInfoManager);
-	static TArray<Worker_ComponentId> CreateClientAuthInterestResultType(USpatialClassInfoManager* ClassInfoManager);
-	static TArray<Worker_ComponentId> CreateServerNonAuthInterestResultType(USpatialClassInfoManager* ClassInfoManager);
-	static TArray<Worker_ComponentId> CreateServerAuthInterestResultType(USpatialClassInfoManager* ClassInfoManager);
+	// TODO: create and pull out into result types class
+	ResultType CreateClientNonAuthInterestResultType(USpatialClassInfoManager* ClassInfoManager);
+	ResultType CreateClientAuthInterestResultType(USpatialClassInfoManager* ClassInfoManager);
+	ResultType CreateServerNonAuthInterestResultType(USpatialClassInfoManager* ClassInfoManager);
+	ResultType CreateServerAuthInterestResultType();
 
-	Interest CreateInterest() const;
+	Interest CreateInterest(AActor* InActor, const FClassInfo& InInfo, const Worker_EntityId InEntityId) const;
 
-	// Only uses Defined Constraint
-	void AddActorInterest(Interest& OutInterest) const;
 	// Defined Constraint AND Level Constraint
-	void AddPlayerControllerActorInterest(Interest& OutInterest) const;
+	void AddPlayerControllerActorInterest(Interest& OutInterest, const AActor* InActor, const FClassInfo& InInfo) const;
+	// Self interests require the entity ID to know which entity is "self". This would no longer be required if there was a first class self constraint.
 	// The components clients need to see on entities they are have authority over that they don't already see through authority.
-	void AddClientSelfInterest(Interest& OutInterest) const;
+	void AddClientSelfInterest(Interest& OutInterest, const Worker_EntityId& EntityId) const;
 	// The components servers need to see on entities they have authority over that they don't already see through authority.
-	void AddServerSelfInterest(Interest& OutInterest) const;
+	void AddServerSelfInterest(Interest& OutInterest, const Worker_EntityId& EntityId) const;
 
-	void GetActorUserDefinedQueries(const AActor* InActor, const QueryConstraint& LevelConstraints, TArray<SpatialGDK::Query>& OutQueries, bool bRecurseChildren) const;
-	TArray<Query> GetUserDefinedQueries(const QueryConstraint& LevelConstraints) const;
+	// Add the checkout radius, always relevant, or always interested query.
+	void AddSystemQuery(Interest& OutInterest, const AActor* InActor, const FClassInfo& InInfo, const QueryConstraint& LevelConstraint) const;
 
-	static void AddComponentQueryPairToInterestComponent(Interest& OutInterest, const Worker_ComponentId ComponentId, const Query& QueryToAdd);
+	void AddUserDefinedQueries(Interest& OutInterest, const AActor* InActor, const QueryConstraint& LevelConstraint) const;
+	FrequencyToConstraintsMap GetUserDefinedFrequencyToConstraintsMap(const AActor* InActor) const;
+	void GetActorUserDefinedQueryConstraints(const AActor* InActor, FrequencyToConstraintsMap& OutFrequencyToConstraints, bool bRecurseChildren) const;
 
-	// Checkout Constraint OR AlwaysInterested OR AlwaysRelevant Constraint
-	QueryConstraint CreateSystemDefinedConstraints() const;
+	void AddNetCullDistanceFrequencyQueries(Interest& OutInterest, const QueryConstraint& LevelConstraint) const;
+
+	void AddComponentQueryPairToInterestComponent(Interest& OutInterest, const Worker_ComponentId ComponentId, const Query& QueryToAdd) const;
 
 	// System Defined Constraints
-	QueryConstraint CreateCheckoutRadiusConstraints() const;
-	QueryConstraint CreateAlwaysInterestedConstraint() const;
-	static QueryConstraint CreateAlwaysRelevantConstraint();
+	QueryConstraint CreateCheckoutRadiusConstraints(const AActor* InActor) const;
+	QueryConstraint CreateAlwaysInterestedConstraint(const AActor* InActor, const FClassInfo& InInfo) const;
+	QueryConstraint CreateAlwaysRelevantConstraint() const;
 
 	// Only checkout entities that are in loaded sub-levels
-	QueryConstraint CreateLevelConstraints() const;	
+	QueryConstraint CreateLevelConstraints(const AActor* InActor) const;
 
 	void AddObjectToConstraint(UObjectPropertyBase* Property, uint8* Data, QueryConstraint& OutConstraint) const;
 
-	AActor* Actor;
-	const FClassInfo& Info;
-	const Worker_EntityId EntityId;
+	// If the result types flag is flipped, set the specified result type.
+	void SetResultType(Query& OutQuery, const ResultType& InResultType) const;
+
+	struct FrequencyConstraint
+	{
+		float Frequency;
+		SpatialGDK::QueryConstraint Constraint;
+	};
+
 	USpatialClassInfoManager* ClassInfoManager;
 	USpatialPackageMapClient* PackageMap;
+
+	// Used to cache checkout radius constraints with frequency settings, so queries can be quickly recreated.
+	TArray<FrequencyConstraint> CheckoutConstraints;
+
+	// The checkout radius constraint is built once for all actors in CreateCheckoutRadiusConstraint as it is equivalent for all actors.
+	// It is built once per net driver initialization.
+	QueryConstraint ClientCheckoutRadiusConstraint;
+
+	// Cache the result types of queries.
+	ResultType ClientNonAuthInterestResultType;
+	ResultType ClientAuthInterestResultType;
+	ResultType ServerNonAuthInterestResultType;
+	ResultType ServerAuthInterestResultType;
 };
 
 } // namespace SpatialGDK